+++ python/bindings/openravepy_global.cpp
@@ -1048,6 +1048,19 @@ BOOST_PYTHON_FUNCTION_OVERLOADS(RetimeAffineTrajectory_overloads, planningutils:
 
 void init_openravepy_global()
 {
+    enum_<OpenRAVEErrorCode>("ErrorCode" DOXY_ENUM(OpenRAVEErrorCode))
+    .value("Failed",ORE_Failed)
+    .value("InvalidArguments",ORE_InvalidArguments)
+    .value("EnvironmentNotLocked",ORE_EnvironmentNotLocked)
+    .value("CommandNotSupported",ORE_CommandNotSupported)
+    .value("Assert",ORE_Assert)
+    .value("InvalidPlugin",ORE_InvalidPlugin)
+    .value("InvalidInterfaceHash",ORE_InvalidInterfaceHash)
+    .value("NotImplemented",ORE_NotImplemented)
+    .value("InconsistentConstraints",ORE_InconsistentConstraints)
+    .value("NotInitialized",ORE_NotInitialized)
+    .value("InvalidState",ORE_InvalidState)
+    ;
     enum_<DebugLevel>("DebugLevel" DOXY_ENUM(DebugLevel))
     .value("Fatal",Level_Fatal)
     .value("Error",Level_Error)